fix: deleteHabit 에서 id를 int에서 long으로 변경

+++ data/src/main/java/com/depromeet/threedays/data/api/HabitService.kt
@@ -29,6 +29,6 @@ interface HabitService {
 
     @DELETE("/api/v1/habits/{habitId}")
     suspend fun deleteHabit(
-        @Path("productId") productId: Int
+        @Path("productId") productId: Long
     )
 }
