+++ domain/src/main/java/com/depromeet/threedays/domain/repository/HabitRepository.kt
@@ -11,5 +11,5 @@ interface HabitRepository {
     suspend fun getArchivedHabits(): Flow<DataState<List<Habit>>>
     //suspend fun getHabit(habitId: Long): Habit
     suspend fun updateHabit(habitId: Long, habit: CreateHabit)
-    suspend fun deleteHabit(habitId: Int): Flow<DataState<Any>>
+    suspend fun deleteHabit(habitId: Long): Flow<DataState<Any>>
 }
